+++ DTO/Theater/MovieDTO.cs
@@ -8,7 +8,7 @@ namespace DTO.Theater
     {
         public class GenreResponse
         {
-            public int GenreId { get; set; }
+            public byte GenreId { get; set; }
             public string GenreName { get; set; } = "";
         }
 
@@ -34,13 +34,13 @@ namespace DTO.Theater
             public string? BannerUrl { get; set; }
             public string? TrailerUrl { get; set; }
             public decimal? ImdbRating { get; set; }
-            public List<int>? GenreIds { get; set; }
+            public List<byte>? GenreIds { get; set; }
         }
 
         public class MovieDetailResponse
         {
             public MovieResponse Movie { get; set; } = new();
-            public List<int> GenreIds { get; set; } = new();
+            public List<byte> GenreIds { get; set; } = new();
             public List<string> Genres { get; set; } = new();
         }
 
@@ -63,7 +63,7 @@ namespace DTO.Theater
             public string? BannerUrl { get; set; }
             public string? TrailerUrl { get; set; }
             public decimal? ImdbRating { get; set; }
-            public List<int>? GenreIds { get; set; }
+            public List<byte>? GenreIds { get; set; }
         }
 
     }
